// src/modules/system/interfaces/bootstrap.interfaces.ts
// src/modules/system/interfaces/bootstrap.interfaces.ts

export interface BootstrapUserDto {
    email: string;
    password: string;
    firstName: string;
    lastName: string;
    licenseNumber?: string;
}

export interface BootstrapResult {
    success: boolean;
    message: string;
    user: {
        id: string;
        email: string;
        globalRole: string;
    };
    bootstrapCompletedAt: Date;
}

export const BOOTSTRAP_COMPLETED_KEY = 'BOOTSTRAP_COMPLETED';